+++ hc/accounts/models.py
@@ -282,8 +282,15 @@ class Project(models.Model):
                 break
         return status
 
-    def have_broken_channels(self):
-        return self.channel_set.exclude(last_error="").exists()
+    def have_channel_issues(self):
+        errors = list(self.channel_set.values_list("last_error", flat=True))
+
+        # It's a problem if a project has no integrations at all
+        if len(errors) == 0:
+            return True
+
+        # It's a problem if any integration has a logged error
+        return True if max(errors) else False
 
 
 class Member(models.Model):